requirements.txt
chromakey==1.0.0
llvmlite==0.35.0
numba==0.52.0
numpy==1.20.0rc2
opencv-python==4.5.1.48
Pillow==8.1.0
PyMatting==1.1.2
scipy==1.6.0
